BIBLECAST#2 - Refactored BibleSearch activity separating out private classes.
