+++ app/src/androidTest/java/au/id/neasbey/biblecast/API/BiblesOrg/BibleAPIResponseParserBiblesOrgTest.java
@@ -13,7 +13,7 @@ import java.util.List;
 
 import au.id.neasbey.biblecast.API.BibleAPIResponse;
 import au.id.neasbey.biblecast.API.BibleAPIResponseParser;
-import au.id.neasbey.biblecast.BiblecastException;
+import au.id.neasbey.biblecast.exception.BiblecastException;
 import au.id.neasbey.biblecast.model.BibleVersion;
 import au.id.neasbey.biblecast.util.UIUtils;
