+++ app/src/main/java/au/id/neasbey/biblecast/API/BibleAPIResponseParser.java
@@ -4,7 +4,7 @@ import android.text.Spanned;
 
 import java.util.List;
 
-import au.id.neasbey.biblecast.BiblecastException;
+import au.id.neasbey.biblecast.exception.BiblecastException;
 import au.id.neasbey.biblecast.model.BibleVersion;
 
 /**
