+++ app/src/main/java/au/id/neasbey/biblecast/GUIHelper/SearchOnSuggestionListener.java
@@ -1,4 +1,4 @@
-package au.id.neasbey.biblecast;
+package au.id.neasbey.biblecast.GUIHelper;
 
 import android.database.Cursor;
 import android.widget.SearchView;
