+++ app/src/main/java/au/id/neasbey/biblecast/GUIHelper/VersionOnItemSelectedListener.java
@@ -1,4 +1,4 @@
-package au.id.neasbey.biblecast;
+package au.id.neasbey.biblecast.GUIHelper;
 
 import android.util.Log;
 import android.view.View;
@@ -6,18 +6,19 @@ import android.widget.AdapterView;
 
 import java.util.List;
 
+import au.id.neasbey.biblecast.BibleSearch;
 import au.id.neasbey.biblecast.model.BibleVersion;
 
 /**
  * Created by craigneasbey on 1/10/15.
  */
-public class SearchOnItemSelectedListener implements AdapterView.OnItemSelectedListener {
+public class VersionOnItemSelectedListener implements AdapterView.OnItemSelectedListener {
 
-    private static final String TAG = SearchOnItemSelectedListener.class.getSimpleName();
+    private static final String TAG = VersionOnItemSelectedListener.class.getSimpleName();
 
     private BibleSearch bibleSearch;
 
-    public SearchOnItemSelectedListener(BibleSearch bibleSearch) {
+    public VersionOnItemSelectedListener(BibleSearch bibleSearch) {
         this.bibleSearch = bibleSearch;
     }
